+++ petitparser/test/reflection_test.dart
@@ -2,56 +2,254 @@ import 'package:petitparser/petitparser.dart';
 import 'package:petitparser/reflection.dart';
 import 'package:test/test.dart';
 
-void expectAll(Iterable<Parser> parsers, Iterable<String> inputs) {
-  final matchedParsers = <Parser, int>{for (final parser in parsers) parser: 0};
-  final matchedInputs = <String, int>{for (final input in inputs) input: 0};
-  for (final parser in parsers) {
-    final tester = parser.end();
-    for (final input in inputs) {
-      if (tester.accept(input)) {
-        matchedInputs[input] = matchedInputs[input]! + 1;
-        matchedParsers[parser] = matchedParsers[parser]! + 1;
-      }
-    }
-  }
-  matchedInputs.forEach((input, count) =>
-      expect(count, greaterThan(0), reason: '$input expected'));
-  matchedParsers.forEach((parser, count) =>
-      expect(count, greaterThan(0), reason: '$parser not expected'));
+// Güting, Erwig, Übersetzerbau, Springer (p.63)
+Map<Symbol, Parser> createUebersetzerbau() {
+  final grammar = <Symbol, Parser>{};
+  grammar[#a] = char('a');
+  grammar[#b] = char('b');
+  grammar[#c] = char('c');
+  grammar[#d] = char('d');
+  grammar[#e] = epsilon();
+  grammar[#B] = grammar[#b]! | grammar[#e]!;
+  grammar[#A] = grammar[#a]! | grammar[#B]!;
+  grammar[#S] = grammar[#A]! & grammar[#B]! & grammar[#c]! & grammar[#d]!;
+  return grammar;
+}
+
+// The canonical grammar to exercise first- and follow-set calculation,
+// likely originally from the dragon-book.
+Map<Symbol, Parser> createDragon() {
+  final grammar = <Symbol, SettableParser>{
+    for (final symbol in [#E, #Ep, #T, #Tp, #F]) symbol: undefined(),
+  };
+  grammar[#E]!.set(grammar[#T]! & grammar[#Ep]!);
+  grammar[#Ep]!.set((char('+') & grammar[#T]! & grammar[#Ep]!).optional());
+  grammar[#T]!.set(grammar[#F]! & grammar[#Tp]!);
+  grammar[#Tp]!.set((char('*') & grammar[#F]! & grammar[#Tp]!).optional());
+  grammar[#F]!.set((char('(') & grammar[#E]! & char(')')) | char('i'));
+  return grammar;
+}
+
+void expectTerminals(Iterable<Parser> parsers, Iterable<String> inputs) {
+  final expectedInputs = {...inputs};
+  final actualInputs = {
+    for (final parser in [for (final parser in parsers) parser.end()])
+      for (final character in [
+        for (var code = 32; code <= 126; code++) String.fromCharCode(code),
+        '',
+      ])
+        if (parser.accept(character)) character
+  };
+  expect(actualInputs, expectedInputs);
 }
 
 // ignore_for_file: deprecated_member_use_from_same_package
 void main() {
-  group('first-set', () {
-    test('plus', () {
-      final parser = char('a').plus();
-      expectAll(firstSet(parser), ['a']);
+  group('analyzer', () {
+    group('isNullable', () {
+      test('plus', () {
+        final parser = char('a').plus();
+        final analyzer = Analyzer(parser);
+        expect(analyzer.isNullable(parser), isFalse);
+      });
+      test('star', () {
+        final parser = char('a').star();
+        final analyzer = Analyzer(parser);
+        expect(analyzer.isNullable(parser), isTrue);
+      });
+      test('optional', () {
+        final parser = char('a').optional();
+        final analyzer = Analyzer(parser);
+        expect(analyzer.isNullable(parser), isTrue);
+      });
+      test('choice', () {
+        final parser = char('a').or(char('b'));
+        final analyzer = Analyzer(parser);
+        expect(analyzer.isNullable(parser), isFalse);
+      });
+      test('epsilon choice', () {
+        final parser = char('a').or(epsilon());
+        final analyzer = Analyzer(parser);
+        expect(analyzer.isNullable(parser), isTrue);
+      });
+      test('sequence', () {
+        final parser = char('a').seq(char('b'));
+        final analyzer = Analyzer(parser);
+        expect(analyzer.isNullable(parser), isFalse);
+      });
+      test('epsilon sequence', () {
+        final parser = epsilon().seq(char('a'));
+        final analyzer = Analyzer(parser);
+        expect(analyzer.isNullable(parser), isFalse);
+      });
+      test('optional sequence', () {
+        final parser = char('a').optional().seq(char('b'));
+        final analyzer = Analyzer(parser);
+        expect(analyzer.isNullable(parser), isFalse);
+      });
+      test('übersetzerbau grammar', () {
+        final parsers = createUebersetzerbau();
+        final analyzer = Analyzer(parsers[#S]!);
+        expect(analyzer.isNullable(parsers[#S]!), isFalse);
+        expect(analyzer.isNullable(parsers[#A]!), isTrue);
+        expect(analyzer.isNullable(parsers[#B]!), isTrue);
+        expect(analyzer.isNullable(parsers[#a]!), isFalse);
+        expect(analyzer.isNullable(parsers[#b]!), isFalse);
+        expect(analyzer.isNullable(parsers[#c]!), isFalse);
+        expect(analyzer.isNullable(parsers[#d]!), isFalse);
+        expect(analyzer.isNullable(parsers[#e]!), isTrue);
+      });
+      test('dragon grammar', () {
+        final parsers = createDragon();
+        final analyzer = Analyzer(parsers[#E]!);
+        expect(analyzer.isNullable(parsers[#E]!), isFalse);
+        expect(analyzer.isNullable(parsers[#Ep]!), isTrue);
+        expect(analyzer.isNullable(parsers[#T]!), isFalse);
+        expect(analyzer.isNullable(parsers[#Tp]!), isTrue);
+        expect(analyzer.isNullable(parsers[#F]!), isFalse);
+      });
     });
-    test('star', () {
-      final parser = char('a').star();
-      expectAll(firstSet(parser), ['a', '']);
+    group('first-set', () {
+      test('plus', () {
+        final parser = char('a').plus();
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.firstSet(parser), ['a']);
+      });
+      test('star', () {
+        final parser = char('a').star();
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.firstSet(parser), ['a', '']);
+      });
+      test('optional', () {
+        final parser = char('a').optional();
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.firstSet(parser), ['a', '']);
+      });
+      test('choice', () {
+        final parser = char('a').or(char('b'));
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.firstSet(parser), ['a', 'b']);
+      });
+      test('epsilon choice', () {
+        final parser = char('a').or(epsilon());
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.firstSet(parser), ['a', '']);
+      });
+      test('sequence', () {
+        final parser = char('a').seq(char('b'));
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.firstSet(parser), ['a']);
+      });
+      test('epsilon sequence', () {
+        final parser = epsilon().seq(char('a'));
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.firstSet(parser), ['a']);
+      });
+      test('optional sequence', () {
+        final parser = char('a').optional().seq(char('b'));
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.firstSet(parser), ['a', 'b']);
+      });
+      test('übersetzerbau grammar', () {
+        final parsers = createUebersetzerbau();
+        final analyzer = Analyzer(parsers[#S]!);
+        expectTerminals(analyzer.firstSet(parsers[#S]!), ['a', 'b', 'c']);
+        expectTerminals(analyzer.firstSet(parsers[#A]!), ['a', 'b', '']);
+        expectTerminals(analyzer.firstSet(parsers[#B]!), ['b', '']);
+        expectTerminals(analyzer.firstSet(parsers[#a]!), ['a']);
+        expectTerminals(analyzer.firstSet(parsers[#b]!), ['b']);
+        expectTerminals(analyzer.firstSet(parsers[#c]!), ['c']);
+        expectTerminals(analyzer.firstSet(parsers[#d]!), ['d']);
+        expectTerminals(analyzer.firstSet(parsers[#e]!), ['']);
+      });
+      test('dragon grammar', () {
+        final parsers = createDragon();
+        final analyzer = Analyzer(parsers[#E]!);
+        expectTerminals(analyzer.firstSet(parsers[#E]!), ['(', 'i']);
+        expectTerminals(analyzer.firstSet(parsers[#Ep]!), ['+', '']);
+        expectTerminals(analyzer.firstSet(parsers[#T]!), ['(', 'i']);
+        expectTerminals(analyzer.firstSet(parsers[#Tp]!), ['*', '']);
+        expectTerminals(analyzer.firstSet(parsers[#F]!), ['(', 'i']);
+      });
     });
-    test('optional', () {
-      final parser = char('a').optional();
-      expectAll(firstSet(parser), ['a', '']);
-    });
-    test('choice', () {
-      final parser = char('a').or(char('b'));
-      expectAll(firstSet(parser), ['a', 'b']);
-    });
-    test('sequence', () {
-      final parser = char('a').seq(char('b'));
-      expectAll(firstSet(parser), ['a']);
-    });
-    test('epsilon sequence', () {
-      final parser = epsilon().seq(char('a'));
-      expectAll(firstSet(parser), ['a']);
-    });
-    test('optional sequence', () {
-      final parser = char('a').optional().seq(char('b'));
-      expectAll(firstSet(parser), ['a', 'b']);
+    group('follow-set', () {
+      test('plus', () {
+        final parser = char('a').plus();
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.followSet(parser), ['']);
+        expectTerminals(analyzer.followSet(parser.children[0]), ['a', '']);
+      });
+      test('star', () {
+        final parser = char('a').star();
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.followSet(parser), ['']);
+        expectTerminals(analyzer.followSet(parser.children[0]), ['a', '']);
+      });
+      test('optional', () {
+        final parser = char('a').optional();
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.followSet(parser), ['']);
+        expectTerminals(analyzer.followSet(parser.children[0]), ['']);
+      });
+      test('choice', () {
+        final parser = char('a').or(char('b'));
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.followSet(parser), ['']);
+        expectTerminals(analyzer.followSet(parser.children[0]), ['']);
+        expectTerminals(analyzer.followSet(parser.children[1]), ['']);
+      });
+      test('epsilon choice', () {
+        final parser = char('a').or(epsilon());
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.followSet(parser), ['']);
+        expectTerminals(analyzer.followSet(parser.children[0]), ['']);
+        expectTerminals(analyzer.followSet(parser.children[1]), ['']);
+      });
+      test('sequence', () {
+        final parser = char('a').seq(char('b'));
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.followSet(parser), ['']);
+        expectTerminals(analyzer.followSet(parser.children[0]), ['b']);
+        expectTerminals(analyzer.followSet(parser.children[1]), ['']);
+      });
+      test('epsilon sequence', () {
+        final parser = epsilon().seq(char('a'));
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.followSet(parser), ['']);
+        expectTerminals(analyzer.followSet(parser.children[0]), ['a']);
+        expectTerminals(analyzer.followSet(parser.children[1]), ['']);
+      });
+      test('optional sequence', () {
+        final parser = char('a').seq(char('b').optional());
+        final analyzer = Analyzer(parser);
+        expectTerminals(analyzer.followSet(parser), ['']);
+        expectTerminals(analyzer.followSet(parser.children[0]), ['b', '']);
+        expectTerminals(analyzer.followSet(parser.children[1]), ['']);
+      });
+      test('übersetzerbau grammar', () {
+        final parsers = createUebersetzerbau();
+        final analyzer = Analyzer(parsers[#S]!);
+        expectTerminals(analyzer.followSet(parsers[#S]!), ['']);
+        expectTerminals(analyzer.followSet(parsers[#A]!), ['b', 'c']);
+        expectTerminals(analyzer.followSet(parsers[#B]!), ['b', 'c']);
+        expectTerminals(analyzer.followSet(parsers[#a]!), ['b', 'c']);
+        expectTerminals(analyzer.followSet(parsers[#b]!), ['b', 'c']);
+        expectTerminals(analyzer.followSet(parsers[#c]!), ['d']);
+        expectTerminals(analyzer.followSet(parsers[#d]!), ['']);
+        expectTerminals(analyzer.followSet(parsers[#e]!), ['b', 'c']);
+      });
+      test('dragon grammar', () {
+        final parsers = createDragon();
+        final analyzer = Analyzer(parsers[#E]!);
+        expectTerminals(analyzer.followSet(parsers[#E]!), [')', '']);
+        expectTerminals(analyzer.followSet(parsers[#Ep]!), [')', '']);
+        expectTerminals(analyzer.followSet(parsers[#T]!), [')', '+', '']);
+        expectTerminals(analyzer.followSet(parsers[#Tp]!), [')', '+', '']);
+        expectTerminals(analyzer.followSet(parsers[#F]!), [')', '+', '*', '']);
+      });
     });
   });
+
   group('iterable', () {
     test('single', () {
       final parser1 = lowercase();
@@ -95,42 +293,42 @@ void main() {
       expect(parsers, [parser1, parser2, parser3]);
     });
   });
-  group('queries', () {
-    group('isNullable', () {
-      test('true', () {
-        expect(isNullable(char('a').optional()), isTrue);
-        expect(isNullable(char('a').optionalWith('b')), isTrue);
-        expect(isNullable(char('a').star()), isTrue);
-        expect(isNullable(char('a').starGreedy(char('b'))), isTrue);
-        expect(isNullable(char('a').starLazy(char('b'))), isTrue);
-        expect(isNullable(epsilon()), isTrue);
-      });
-      test('false', () {
-        expect(isNullable(char('a')), isFalse);
-        expect(isNullable(char('a').and()), isFalse);
-        expect(isNullable(char('a').not()), isFalse);
-        expect(isNullable(char('a').or(char('b'))), isFalse);
-        expect(isNullable(char('a').plus()), isFalse);
-        expect(isNullable(char('a').seq(char('b'))), isFalse);
-        expect(isNullable(failure()), isFalse);
-      });
-    });
-    group('isTerminal', () {
-      test('true', () {
-        expect(isTerminal(char('a')), isTrue);
-        expect(isTerminal(epsilon()), isTrue);
-        expect(isTerminal(failure()), isTrue);
-        expect(isTerminal(string('a')), isTrue);
-      });
-      test('false', () {
-        expect(isTerminal(char('a').and()), isFalse);
-        expect(isTerminal(char('a').not()), isFalse);
-        expect(isTerminal(char('a').or(char('b'))), isFalse);
-        expect(isTerminal(char('a').plus()), isFalse);
-        expect(isTerminal(char('a').seq(char('b'))), isFalse);
-      });
-    });
-  });
+  // group('queries', () {
+  //   group('isNullable', () {
+  //     test('true', () {
+  //       expect(isNullable(char('a').optional()), isTrue);
+  //       expect(isNullable(char('a').optionalWith('b')), isTrue);
+  //       expect(isNullable(char('a').star()), isTrue);
+  //       expect(isNullable(char('a').starGreedy(char('b'))), isTrue);
+  //       expect(isNullable(char('a').starLazy(char('b'))), isTrue);
+  //       expect(isNullable(epsilon()), isTrue);
+  //     });
+  //     test('false', () {
+  //       expect(isNullable(char('a')), isFalse);
+  //       expect(isNullable(char('a').and()), isFalse);
+  //       expect(isNullable(char('a').not()), isFalse);
+  //       expect(isNullable(char('a').or(char('b'))), isFalse);
+  //       expect(isNullable(char('a').plus()), isFalse);
+  //       expect(isNullable(char('a').seq(char('b'))), isFalse);
+  //       expect(isNullable(failure()), isFalse);
+  //     });
+  //   });
+  //   group('isTerminal', () {
+  //     test('true', () {
+  //       expect(isTerminal(char('a')), isTrue);
+  //       expect(isTerminal(epsilon()), isTrue);
+  //       expect(isTerminal(failure()), isTrue);
+  //       expect(isTerminal(string('a')), isTrue);
+  //     });
+  //     test('false', () {
+  //       expect(isTerminal(char('a').and()), isFalse);
+  //       expect(isTerminal(char('a').not()), isFalse);
+  //       expect(isTerminal(char('a').or(char('b'))), isFalse);
+  //       expect(isTerminal(char('a').plus()), isFalse);
+  //       expect(isTerminal(char('a').seq(char('b'))), isFalse);
+  //     });
+  //   });
+  // });
   group('transform', () {
     test('copy', () {
       final input = lowercase().settable();